Add PreprocessorVariable class in MCC compiler
This commit introduced a new PreprocessorVariable class in the MCC compiler. This class, functioning as a container, can hold multiple dynamic values with support for reallocation similar to list behavior when necessary. It incorporates the usability of both an array and a List, and includes functionalities like indexing, setting values, adding and removing elements.