Refactor DefaultPackManager into TemporaryFilesManager class
DefaultPackManager class has been deleted and replaced by TemporaryFilesManager class. This new class offers the similar functionalities of the DefaultPackManager class along with extra methods for file reading and writing from, and to, the temporary directory. This change is designed to centralize and streamline the handling of temporary files and caching.